Move json stuff to json_io.py
