+++ bayem/__init__.py
@@ -1,5 +1,6 @@
 from .distributions import *
 from .visualization import *
+from .json_io import *
 from .vba import *
 
 __version__ = "2021.0.0"
