+++ bayem/vba.py
@@ -1,6 +1,7 @@
 import copy
 import json
 import logging
+from dataclasses import dataclass
 from time import perf_counter
 from tabulate import tabulate
 
@@ -11,6 +12,96 @@ from .distributions import MVN, Gamma
 
 logger = logging.getLogger(__name__)
 
+@dataclass
+class Options:
+    tolerance : float = 0.1
+    maxiter : int = 50
+
+class DictProblem:
+    def __init__(self, f, noise0, jac):
+        self.f = f
+        self.noise0 = noise0
+        self.jac = jac
+
+
+    def __call__(self, x):
+        """
+        Computes (f(x), jac(x)) at the current parameter mean `x` as a dict
+        containing {noise_key: np.ndarray}.
+        """
+        return self.f(x), self.jac(x)
+
+    def first_call(self, x):
+        """
+        Computes (f(x), jac(x)) at the current parameter mean `x` and sets
+        up an internal structure such that various user provided forms of 
+        `f` can be used.
+        """
+        return self(x)
+
+class VBA:
+    def __init__(self, dict_problem, x0, options):
+        self.p = dict_problem
+        self.x0 = x0
+        self.noise0 = self.p.noise0
+        self.options = options
+        self.result = VBResult()
+        pass
+
+    def run(self):
+        t0 = perf_counter()
+        # extract / rename model parameters
+        m = np.array(self.x0.mean)
+        L = np.array(self.x0.precision)
+        m0 = np.copy(m)
+        L0 = np.copy(L)
+
+        # run first evaluation of the f and jac to adjust the format
+        k, J = self.p.first_call(m)
+
+        # extract / rename noise parameters
+        s, c = {}, {}
+        for n, gamma in self.noise0.items():
+            s[n] = gamma.scale
+            c[n] = gamma.shape
+        s0 = copy.copy(s)
+        c0 = copy.copy(c)
+
+
+        pass
+
+
+
+def vba(f, x0, noise0=None, jac=None, options=Options()):
+    """
+    Implementation of 
+        Variational Bayesian inference for a nonlinear 
+        forward model [Chapell et al, 2008]
+    for an arbitrary `model_error` f.
+
+    Parameters
+    ==========
+
+    The implementation is close to the formulas of 
+    section III.C (Extending the Noise Model)
+    with the same notation and references to each formula, with the only
+    exception that capital lambda (precision) in the paper is here referred 
+    to as L.
+
+    @article{chappell2008variational,
+          title={Variational Bayesian inference for a nonlinear forward model},
+          author={Chappell, Michael A and Groves, Adrian R and Whitcher,
+                  Brandon and Woolrich, Mark W},
+          journal={IEEE Transactions on Signal Processing},
+          volume={57},
+          number={1},
+          pages={223--236},
+          year={2008},
+          publisher={IEEE}
+        }
+    """
+    dict_problem = DictProblem(f, noise0, jac)
+    return VBA(dict_problem, x0, options).run()
 
 class VariationalBayesInterface:
     def __call__(self, number_vector):
@@ -458,123 +549,4 @@ class VB:
         return False
 
 
-class BayesEncoder(json.JSONEncoder):
-    """
-    Usage:
-        string = json.dumps(obj, cls=bayes.vb.BayesEncoder, ...)
-        with open(...) as f:
-            json.dump(obj, f cls=bayes.vb.BayesEncoder, ...)
 
-    Details:
-
-    Out of the box, JSON can serialize
-        dict, list, tuple, str, int, float, True/False, None
-
-    To make our custom classes JSON serializable, we subclass from JSONEncoder
-    and overwrite its `default` method to somehow represent our class with the
-    types provided above.
-
-    The idea is to serialize our custom classes (and numpy...) as a dict
-    containing:
-        key: unique string to represent the classe -- this helps us to indentify
-             the class when _de_serializing in `bayes.vb.bayes_hook` below
-        value: some json-serializable entries -- obj.__dict__ contains all
-               members class members and is not optimal, but very convenient.
-    https://stackoverflow.com/questions/3768895/how-to-make-a-class-json-serializable
-
-    Note that the method is called recursively. To serialize MVN we call the
-    `default` method below and see that json representation should contain
-    the __dict__ of members. This dict also contains "mean" of type `np.array`.
-    Thus, `default` is now called on `np.array` which is represented by a list
-    of its values.
-
-    """
-
-    def default(self, obj):
-        """
-        `obj`:
-            python object to serialize
-        return:
-            json (serializeable) reprensentation of `obj`
-        """
-        if isinstance(obj, VBResult):
-            return {"bayem.VBResult": obj.__dict__}
-
-        if isinstance(obj, MVN):
-            return {"bayem.MVN": obj.__dict__}
-
-        if isinstance(obj, Gamma):
-            return {"bayem.Gamma": obj.__dict__}
-
-        if isinstance(obj, np.ndarray):
-            return {"np.array": obj.tolist()}
-
-        # `obj` is not one of our types? Fall back to superclass implementation.
-        return json.JSONEncoder.default(self, obj)
-
-
-def bayes_hook(dct):
-    """
-    `dct`:
-        json reprensentation of an `obj` (a dict)
-    `obj`:
-        python object created from `dct`
-
-    Usage:
-        obj = json.loads(string, object_hook=bayes.vb.bayes_hook, ...)
-        with open(...) as f:
-            obj = json.load(f, object_hook=bayes.vb.bayes_hook, ...)
-
-    Details:
-
-    BayesEncoder stores all of our classes as dicts containing their members.
-    This `object_hook` tries to convert those dicts back to actual python objects.
-
-    Some fancy metaprogramming may help to avoid repetition. TODO?
-    """
-    if "bayem.VBResult" in dct:
-        result = VBResult()
-        result.__dict__ = dct["bayem.VBResult"]
-        return result
-
-    if "bayem.MVN" in dct:
-        mvn = MVN()
-        mvn.__dict__ = dct["bayem.MVN"]
-        return mvn
-
-    if "bayem.Gamma" in dct:
-        gamma = Gamma()
-        gamma.__dict__ = dct["bayem.Gamma"]
-        return gamma
-
-    if "np.array" in dct:
-        return np.array(dct["np.array"])
-
-    # Type not recognized, just return the dict.
-    return dct
-
-
-def save_json(obj, filename=None):
-    """
-    Saves an `obj` (possibly containing VB classes) to `filename` via json or
-    returns the json string.
-    """
-    s = json.dumps(obj, cls=BayesEncoder, indent=2)
-    if filename is not None:
-        with open(filename, "w") as f:
-            f.write(s)
-    return s
-
-
-def load_json(filename_or_string):
-    """
-    Loads an object (possibly containing VB classes) from `filename_or_string` 
-    via json.
-    """
-    if filename_or_string.endswith(".json"):
-        with open(filename_or_string, "r") as f:
-            string = f.read()
-    else:
-        string = filename_or_string
-
-    return json.loads(string, object_hook=bayes_hook)
